FoSBot_Backup/FoSBot/requirements.txt
# Version History: 0.7.2 -> 0.7.3
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
twitchio>=2.10.0
google-api-python-client>=2.80.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
tweepy>=4.13.0
websockets>=11.0.0
httpx>=0.27.0
requests-oauthlib>=1.3.0
aiohttp>=3.8.0
nest-asyncio>=1.5.0
aiofiles>=23.1.0
pydantic>=2.0.0
typing-extensions>=4.8.0
click>=7.0
h11>=0.8
httptools>=0.5.0
pyyaml>=5.1
uvloop>=0.17.0
watchfiles>=0.13
certifi
googleapis-common-protos>=1.56.2,<2.0.0
protobuf>=3.19.5,<7.0.0
proto-plus>=1.22.3,<2.0.0
cachetools>=2.0.0,<6.0
pyasn1-modules>=0.2.1
rsa>=3.1.4,<5
pyparsing!=3.0.0,!=3.0.1,!=3.0.2,!=3.0.3,<4,>=2.4.2
charset-normalizer<4,>=2
urllib3<3,>=1.21.1
sniffio>=1.1
pyasn1>=0.6.1,<0.7.0
attrs>=17.3.0
frozenlist>=1.1.1
multidict<7.0,>=4.5
yarl<2.0,>=1.0
annotated-types>=0.4.0
typing-inspection>=0.4.0
idna
aiosignal>=1.1.2
oauthlib<4,>=3.2.0
requests<3,>=2.27.0
httpcore>=1.0.0,<2.0.0
aiohappyeyeballs>=2.3.0
propcache>=0.2.0
